+++ src/Client.php
@@ -20,7 +20,7 @@ class Client implements ClientContract
 
     private const TIMEOUT = 180;
 
-    private const VERSION = '1.115.0';
+    private const VERSION = '1.116.0';
 
     private const USER_AGENT = 'cyberfusion-cluster-api-client/' . self::VERSION;